phase2/requirements.txt
# Phase 2 Python Dependencies

# Core Dependencies
pyyaml>=6.0

#Testing Dependencies (optional, install with: pip install -e .[dev])
#pytest>=7.0
#pytest-cov>=3.0
